% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_states.R
\name{college_states}
\alias{college_states}
\title{Adds college state information}
\usage{
college_states(recruits_df)
}
\arguments{
\item{recruits_df}{Recruiting dataframe (\emph{DF}, required)}
}
\description{
Adds college state information
}
\examples{

college_states(recruits_df)

}
\keyword{internal}
