% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{recruitR-package}
\alias{recruitR}
\alias{recruitR-package}
\title{recruitR: A College Sports Recruiting Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

As a college sports recruiting package, recruitR provides users with the access to 247 Composite rankings and recruit information from the CollegeFootballData API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://recruitR.sportsdataverse.org/}
  \item \url{https://github.com/sportsdataverse/recruitR}
  \item Report bugs at \url{https://github.com/sportsdataverse/recruitR/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com}

Authors:
\itemize{
  \item Ryan Mann \email{mannrcfe@gmail.com}
}

}
\keyword{internal}
