% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfs_recruiting_rankings.R
\name{tfs_recruiting_rankings}
\alias{tfs_recruiting_rankings}
\title{\strong{Get 247 CFB Player Recruiting Rankings}}
\usage{
tfs_recruiting_rankings(year, recruit_type = "HighSchool", pages = 5)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY}) - Minimum: 2000, Maximum: 2020 currently}

\item{recruit_type}{(\emph{String} optional): default API return is 'HighSchool', other options include 'JUCO'
or 'PrepSchool'  - For position group information}

\item{pages}{Number of pages}
}
\value{
\code{\link[=tfs_recruiting_rankings]{tfs_recruiting_rankings()}}
}
\description{
\strong{Get 247 CFB Player Recruiting Rankings}
}
\examples{
\donttest{
   try(tfs_recruiting_rankings(2022))
}

}
\keyword{Recruiting}
