% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfs_recruitment.R
\name{tfs_recruitment}
\alias{tfs_recruitment}
\title{\strong{Get 247 Recruitment}}
\usage{
tfs_recruitment(recruitment_id)
}
\arguments{
\item{recruitment_id}{(\emph{Integer})}
}
\value{
\code{\link[=tfs_recruitment]{tfs_recruitment()}}
}
\description{
\strong{Get 247 Recruitment}
}
\examples{
\donttest{
   tfs_recruitment(recruitment_id = 130361)
}

}
\keyword{Recruiting}
